+++ sdk/clientcore/http-netty4/src/main/java/io/clientcore/http/netty4/implementation/Netty4InitiateOneReadHandler.java
@@ -9,6 +9,7 @@ import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandler;
 import io.netty.channel.ChannelInboundHandlerAdapter;
 import io.netty.handler.codec.http.LastHttpContent;
+import io.netty.handler.codec.http2.Http2DataFrame;
 import io.netty.util.ReferenceCountUtil;
 
 import java.io.IOException;
@@ -21,6 +22,8 @@ import java.util.concurrent.CountDownLatch;
  */
 public final class Netty4InitiateOneReadHandler extends ChannelInboundHandlerAdapter {
     private final IOExceptionCheckedConsumer<ByteBuf> byteBufConsumer;
+    private final boolean isHttp2;
+
     private CountDownLatch latch;
 
     private boolean lastRead;
@@ -35,10 +38,13 @@ public final class Netty4InitiateOneReadHandler extends ChannelInboundHandlerAda
      *
      * @param latch The latch to count down when the channel read completes.
      * @param byteBufConsumer The consumer to process the {@link ByteBuf ByteBufs} as they are read.
+     * @param isHttp2 Flag indicating whether the handler is used for HTTP/2 or not.
      */
-    public Netty4InitiateOneReadHandler(CountDownLatch latch, IOExceptionCheckedConsumer<ByteBuf> byteBufConsumer) {
+    public Netty4InitiateOneReadHandler(CountDownLatch latch, IOExceptionCheckedConsumer<ByteBuf> byteBufConsumer,
+        boolean isHttp2) {
         this.latch = latch;
         this.byteBufConsumer = byteBufConsumer;
+        this.isHttp2 = isHttp2;
     }
 
     /**
@@ -77,7 +83,11 @@ public final class Netty4InitiateOneReadHandler extends ChannelInboundHandlerAda
             }
         }
 
-        lastRead = msg instanceof LastHttpContent;
+        if (isHttp2) {
+            lastRead = msg instanceof Http2DataFrame && ((Http2DataFrame) msg).isEndStream();
+        } else {
+            lastRead = msg instanceof LastHttpContent;
+        }
         ctx.fireChannelRead(msg);
     }