+++ sdk/clientcore/http-netty4/src/main/java/io/clientcore/http/netty4/implementation/Netty4ResponseHandler.java
@@ -25,8 +25,8 @@ import static io.clientcore.http.netty4.implementation.Netty4Utility.readByteBuf
 import static io.clientcore.http.netty4.implementation.Netty4Utility.setOrSuppressError;
 
 /**
- * A {@link ChannelInboundHandler} implementation that appropriately handles the response reading from the server based
- * on the information provided from the headers.
+ * A {@link ChannelInboundHandler} implementation that appropriately handles {@code HTTP/1.1} responses by using the
+ * response headers to determine how to read the response from the server.
  * <p>
  * When used with {@code NettyHttpClient} this handler must be added to the pipeline so that the {@link HttpClientCodec}
  * is able to decode the data of the response.
@@ -77,7 +77,7 @@ public final class Netty4ResponseHandler extends ChannelInboundHandlerAdapter {
     }
 
     @Override
-    public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) throws Exception {
+    public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) {
         setOrSuppressError(errorReference, cause);
         latch.countDown();
     }
@@ -126,8 +126,8 @@ public final class Netty4ResponseHandler extends ChannelInboundHandlerAdapter {
             started = true;
             HttpResponse response = (HttpResponse) msg;
             this.statusCode = response.status().code();
-            this.headers = (response.headers() instanceof WrappedHttpHeaders)
-                ? ((WrappedHttpHeaders) response.headers()).getCoreHeaders()
+            this.headers = (response.headers() instanceof WrappedHttp11Headers)
+                ? ((WrappedHttp11Headers) response.headers()).getCoreHeaders()
                 : Netty4Utility.convertHeaders(response.headers());
 
             if (msg instanceof FullHttpResponse) {
@@ -173,11 +173,7 @@ public final class Netty4ResponseHandler extends ChannelInboundHandlerAdapter {
         }
 
         responseReference.set(new ResponseStateInfo(ctx.channel(), complete, statusCode, headers, eagerContent,
-            ResponseBodyHandling.getBodyHandling(request, headers)));
+            ResponseBodyHandling.getBodyHandling(request, headers), false));
         latch.countDown();
-    }
-
-    private enum BodyHandling {
-        IGNORE, STREAM, BUFFER
     }
 }