+++ sdk/eventgrid/azure-messaging-eventgrid/src/main/java/com/azure/messaging/eventgrid/systemevents/MediaLiveEventChannelArchiveHeartbeatEventData.java
@@ -16,8 +16,12 @@ import java.time.Duration;
 /**
  * Channel Archive heartbeat event data. Schema of the data property of an EventGridEvent for a
  * Microsoft.Media.LiveEventChannelArchiveHeartbeat event.
+ * 
+ * @deprecated This class is deprecated and may be removed in future releases. System events are now available in the
+ * azure-messaging-eventgrid-systemevents package.
  */
 @Immutable
+@Deprecated
 public final class MediaLiveEventChannelArchiveHeartbeatEventData
     implements JsonSerializable<MediaLiveEventChannelArchiveHeartbeatEventData> {