+++ sdk/schemaregistry/azure-data-schemaregistry-jsonschema/CHANGELOG.md
@@ -10,6 +10,20 @@
 
 ### Other Changes
 
+## 1.0.2 (2025-06-19)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-data-schemaregistry` from `1.5.4` to version `1.6.0-beta.1`.
+- Upgraded `azure-core` from `1.55.3` to version `1.55.4`.
+
+#### Dependency Updates
+
+- Upgraded `azure-data-schemaregistry` from `1.5.4` to version `1.6.0-beta.1`.
+- Upgraded `azure-core` from `1.55.3` to version `1.55.4`.
+
 ## 1.0.1 (2025-03-24)
 
 ### Other Changes